mv CLIRunner into lib.rs
